% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{create_county_boundaries}
\alias{create_county_boundaries}
\title{create_county_boundaries}
\usage{
create_county_boundaries(statey, countyy)
}
\arguments{
\item{statey}{String}

\item{countyy}{String}
}
\value{
SFC
}
\description{
create_county_boundaries
}
\examples{
create_county_boundaries(statey = 'MA', countyy = 'Norfolk')
}
