% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishnetutils.R, R/road-buffer.R
\name{create_fishnet_clip}
\alias{create_fishnet_clip}
\title{create_fishnet_clip}
\usage{
create_fishnet_clip(inputty, crs_to, cell_size = 5280)

create_fishnet_clip(inputty, crs_to, cell_size = 5280)
}
\arguments{
\item{inputty}{SF or SFC}

\item{crs_to}{Int or numeric}

\item{cell_size}{Int or numeric}
}
\value{
SFC of a full land area fishnet

SFC of fishnet points, not clipped
}
\description{
create_fishnet_clip

create_fishnet_clip
}
