% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road-buffer.R
\name{make_buffered_road}
\alias{make_buffered_road}
\title{make_buffered_road}
\usage{
make_buffered_road(state_name, county_name, crs_in, buffer_width)
}
\arguments{
\item{state_name}{Character,}

\item{county_name}{Character,}

\item{crs_in}{Numeric, int, a planar CRS}

\item{buffer_width}{Int numeric, assumes unit of crs_in. Width of the buffer to build around the road}
}
\value{
SFC containing a multipolygon showing a road buffer.
}
\description{
make_buffered_road
}
