% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{save_file}
\alias{save_file}
\title{save_file}
\usage{
save_file(to_save, geo_type, format, county_name, county_abbrev)
}
\arguments{
\item{to_save}{SF or SFC}

\item{geo_type}{Character or string}

\item{format}{Character or string}

\item{county_name}{Character or string}

\item{county_abbrev}{Character or string}
}
\value{
Empty, saves output file
}
\description{
save_file
}
